Extract creation of did doc from services
